app/requirements/base.txt
# Django
django==2.2.16
psycopg2==2.8.5 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
django-braces==1.14.0
python-dateutil==2.8.1
django-filter==2.4.0
django-tables2==2.3.4
django-grappelli==2.14.3
coreapi==2.3.3
openapi-codec==1.3.2
djangorestframework==3.12.4
simplejson==3.17.2
django-rest-auth==0.9.5
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
drf-yasg==1.20.0
#drf-yasg2
pytz>=2021.1
django-allauth==0.44.0
Pillow==8.2.0
django-model-utils==4.1.1
django-registration-redux==2.9
django-extra-fields==3.0.2
drf-writable-nested==0.6.2
django-jsonfield==1.4.1
drf-jwt==1.19.0 #1.17.3
django-import-export==2.5.0
django-jazzmin==2.4.5
django-advanced-filters==1.3.0
django-admin-rangefilter==0.8.1  # https://github.com/silentsokolov/django-admin-rangefilter

#Fuzzy String Matching
fuzzywuzzy[speedup]==0.18.0 # https://github.com/seatgeek/fuzzywuzzy

# XlsxWriter
XlsxWriter==1.3.8
# openpyxl
openpyxl==3.0.7
drf-renderer-xlsx==0.4.0

# Gunicorn
gunicorn==20.0.4  # https://github.com/benoitc/gunicorn

# Serve static file
whitenoise==5.2.0

# Remove unused upload files
django-unused-media==0.2.2
django-cleanup==5.1.0
django-admin-shortcuts==2.0.0

django-protected-media==1.0.0 # https://github.com/cobusc/django-protected-media

# django-useraudit==1.7.1
# django-defender==0.8.0
django-axes==5.14.0